requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
Django==5.1.3
dotenv-python==0.0.1
filelock==3.16.1
fonttools==4.54.1
fpdf==1.7.2
frozenlist==1.5.0
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
markdown2==2.5.1
MarkupSafe==3.0.2
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.3
packaging==24.1
pdfkit==1.0.0
pillow==11.0.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyht==0.1.5
pyparsing==3.2.0
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
setuptools==75.4.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.67.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==6.0.0
websockets==13.1
yarl==1.17.1
youtube-transcript-api==0.6.2
